added hostname on container creating
Why I need it. I need to write integration and end-to-end tests using
Kafka. Kafka needs Zookeeper. When you start them in containers you
need to configure it in some specific way. For that you need to
reference Kafka in container one time with `localhost` and another with
its hostname. If you don't know or don't control this hostname your
Kafka will be unreachable. The easiest way to accomplish this is to
set it during the container creating. Docker Compose for instance allows
to set the container internal hostname (and a Docker network hostname)
in the compose-file. Docker API allows to do it but this library for
some reason ignores this setting. I've added this missing ability.
I've seen a different approach to fix this problem in samples but it
requires some shell-scripts to run during the Kafka container startup.
It's too complex and not flexible. What if you can't (and in most cases
can't) alter your Kafka image? My way to fix this problem is much more
elegant and easy but requires to change some internals of this library.