Update Pulsar image version and extend auth token validity
Pulsar image version in PulsarBuilder has been updated from 3.0.4 to 3.2.3. The authentication token validity period has been extended from 1 hour to 365 days in both PulsarBuilder and PulsarContainerTest to reduce the impact of a token creation bug in Apache Pulsar. This change significantly reduces the impact of a token creation bug that affected versions 3.2.0 to 3.2.3. The bug is related to mishandling time units (seconds as milliseconds).