Update authToken creation to support infinite expiration
This update modifies the CreateAuthenticationTokenAsync function to accept an expiry time, including an option for infinite expiration. In addition, the PulsarContainerTest and Pulsar documentation have been updated accordingly. The function still warns of a known regression issue when image tags start with "3.2" or "latest", converting the expiry time unit improperly.