Refactor KeycloakBuilder and add health tests
This commit makes significant updates to the KeycloakBuilder and introduces a new test class for Keycloak container health. In the KeycloakBuilder, the Build method has been revised to remove the immediate call to `Validate()` and the direct return of a new `KeycloakContainer`. Instead, it now includes logic to extract the major version number from the `DockerResourceConfiguration.Image.Tag`. If the version is 25 or higher, the builder is modified to add port binding for the Keycloak health port and a wait strategy for the `/health/ready` endpoint. After these potential modifications, `Validate()` is called to ensure the configuration is correct before returning the container.

Additionally, the validation logic that previously checked the Docker image tag's major version has been removed from the direct validation method and incorporated into the Build method. This change streamlines the validation process, focusing it solely on configuration validation.

A new test class, `KeycloakContainerHealthTest`, has been added to verify the health of the Keycloak container. It includes a method `HealthIsReady` that tests whether the container starts successfully and if its health endpoint returns an HTTP OK status. This method dynamically adjusts the `HttpClient.BaseAddress` based on the container's accessible hostname and port, ensuring the test is robust across different environments. The test covers both below and at the threshold versions with inline data for "quay.io/keycloak/keycloak:24.0" and "quay.io/keycloak/keycloak:25.0", ensuring the new health check configuration is properly applied. Containers are disposed of post-test to prevent resource leakage.