chore: Update Pulsar version and improve affected version check
The default Pulsar image version changed from "3.2.3" to "3.0.6" due to a known regression impacting versions "3.2.0-3.2.3" and "3.3.0". The regression resulted in the expiry time being converted to an incorrect unit of time.
The revision improves the check for affected versions by introducing the 'IsVersionAffectedByRegression' method, which accepts a version and returns true if the version is impacted by the regression.
The formatting of the PulsarContainer.cs and PulsarBuilder.cs files was also updated for better readability.
In addition, the `docs/modules/index.md` file was updated to reflect the new Pulsar version, making it easier for users to use the correct version.