Throw the new `DockerUnavailableException` when Docker is not available
(because it is either not running or misconfigured)

The `ArgumentException` was not the appropriate exception to throw since the root of the problem is not a bad argument but the state of the system which is wrong.

Also improve the error message to mention all the endpoints are applicable but not available.