fix: Reflect Unix file mode when building tarball
When the image build process is creating the tarball of files needed to
build the image, it currently just uses the default value of 700. This
causes those files to be readable only by the owner (i.e. root) when
placed in the container.

The fix is to read the file mode of the files when building the tarball
on Unix. The API in use only exists in .NET 7.0 and higher, so this
change also extends the Windows workaround to Unix. When the
`GetUnixFileMode` function isn't available, just use 755 as the file
mode.